Czyszczenie śmieci

+++ STM_Round/Core/Src/SX1278.c
@@ -82,7 +82,7 @@ void SX1278_SPIWrite(SX1278_t *module, uint8_t addr, uint8_t cmd) {
 }
 
 void SX1278_SPIBurstRead(SX1278_t *module, uint8_t addr, uint8_t *rxBuf,
-		uint8_t length) {
+		uint16_t length) {
 	uint8_t i;
 	if (length <= 1) {
 		return;
@@ -97,7 +97,7 @@ void SX1278_SPIBurstRead(SX1278_t *module, uint8_t addr, uint8_t *rxBuf,
 }
 
 void SX1278_SPIBurstWrite(SX1278_t *module, uint8_t addr, uint8_t *txBuf,
-		uint8_t length) {
+		uint16_t length) {
 	unsigned char i;
 	if (length <= 1) {
 		return;
@@ -192,7 +192,7 @@ void SX1278_clearLoRaIrq(SX1278_t *module) {
 	SX1278_SPIWrite(module, LR_RegIrqFlags, 0xFF);
 }
 
-int SX1278_LoRaEntryRx(SX1278_t *module, uint8_t length, uint32_t timeout) {
+int SX1278_LoRaEntryRx(SX1278_t *module, uint16_t length, uint32_t timeout) {
 	uint8_t addr;
 
 	module->packetLength = length;
@@ -247,7 +247,7 @@ uint8_t SX1278_LoRaRxPacket(SX1278_t *module) {
 	return module->readBytes;
 }
 
-int SX1278_LoRaEntryTx(SX1278_t *module, uint8_t length, uint32_t timeout) {
+int SX1278_LoRaEntryTx(SX1278_t *module, uint16_t length, uint32_t timeout) {
 	uint8_t addr;
 	uint8_t temp;
 
@@ -278,7 +278,7 @@ int SX1278_LoRaEntryTx(SX1278_t *module, uint8_t length, uint32_t timeout) {
 	}
 }
 
-int SX1278_LoRaTxPacket(SX1278_t *module, uint8_t *txBuffer, uint8_t length,
+int SX1278_LoRaTxPacket(SX1278_t *module, uint8_t *txBuffer, uint16_t length,
 		uint32_t timeout) {
 	SX1278_SPIBurstWrite(module, 0x00, txBuffer, length);
 	SX1278_SPIWrite(module, LR_RegOpMode, 0x8b);	//Tx Mode
@@ -310,7 +310,7 @@ void SX1278_begin(SX1278_t *module, uint64_t frequency, uint8_t power,
 	SX1278_defaultConfig(module);
 }
 
-int SX1278_transmit(SX1278_t *module, uint8_t *txBuf, uint8_t length,
+int SX1278_transmit(SX1278_t *module, uint8_t *txBuf, uint16_t length,
 		uint32_t timeout) {
 	if (SX1278_LoRaEntryTx(module, length, timeout)) {
 		return SX1278_LoRaTxPacket(module, txBuf, length, timeout);
@@ -318,7 +318,7 @@ int SX1278_transmit(SX1278_t *module, uint8_t *txBuf, uint8_t length,
 	return 0;
 }
 
-int SX1278_receive(SX1278_t *module, uint8_t length, uint32_t timeout) {
+int SX1278_receive(SX1278_t *module, uint16_t length, uint32_t timeout) {
 	return SX1278_LoRaEntryRx(module, length, timeout);
 }
 
@@ -326,7 +326,7 @@ uint8_t SX1278_available(SX1278_t *module) {
 	return SX1278_LoRaRxPacket(module);
 }
 
-uint8_t SX1278_read(SX1278_t *module, uint8_t *rxBuf, uint8_t length) {
+uint8_t SX1278_read(SX1278_t *module, uint8_t *rxBuf, uint16_t length) {
 	if (length != module->readBytes)
 		length = module->readBytes;
 	memcpy(rxBuf, module->rxBuffer, length);
